refactor tests for Iptu payment
